PLUS-trunk: =Password Encryption Revisited= k2r released a nice little password encryption mod for the masses, but it had a couple of problems:
==Original Release==
 1. New user managers (i.e. SQL, multi-master, etc - if/when they are ever made available) would have to be coded with specific support for the encryption mod for those users who are using it, and coded a different way for users who are not using it.
 2. It advertises MD5 passwords but only provides unix style crypt.
 3. The crypt support is implemented in a flawed manner.  The first two characters of a user's password are used as the salt given to crypt, which is then also stored in the userfile in plain text, reducing the brute force necessary to crack each password should it be compromised.  Only the first 8 characters are significant when using the crypt algorithm, giving away 2 characters compromises at minimum 25% of everyones passwords.
 4. While it provides crypted passwords for new users, it does nothing at all for existing users.

==Second Release==
In this version, which I would like to call an update to his mod (hopefully with his blessing):
 1. A minimalist approach was used to make this compatibile with future user managers, w/o the need for those usermanagers to provide any direct support of this mod.
 2. The fabled MD5 hashing has been added
 3. The crypt method from the last version has been fixed to generate salts at random.
 4. A new configuration option is provided in drftpd.conf.dist, giving you the ability to choose between md5, crypt, or even no encryption at all.
 5. All of your users will be converted to your "preferred" format over time (conversion is done when they log in)  You can convert from any format to any format at any time without even having to restart.  You only need to change the config and SITE RELOAD.

''Be advised, if you do not configure it, it will convert all existing users to the '''MD5''' hash password format.''

==Release Note==
The MD5 format used by this mod, is completely compatible with Linux/BSD password files, to address zubov's concerns of portability.  The crypt format is merely prefixed with a + which is easy to strip off (this +, while not ideal is kept to maintain backwards compatibility).  Should you ever need yet another format, you can just change the config option and convert all users back to unencrypted on very little notice before you need to convert to something else.

Also, be aware that drftpd does not actually SAVE users to disk as they are modified so you can not see the effect of switching back and forth between encryption types very easily unless you shut down your master in between, for this reason I have provided debug logging that shows you when a user is being converted, and the from/to types.

==Known Bugs==
If a users password is stored unencrypted and starts with either '+' or '$1$' then the software will get confused into thinking that it actually is encrypted.  If you are not using encryption at all, and you are using this mod a user could lock themselves by choosing such a password.  As far as I know, this is in no way exploitable.  This bug is present in k2r's version as well.

==TODO==
I'd like to one-day experiment with storing the password as a Password, PasswordMD5, or PasswordCrypt object (or a Password object that holds both the password and the type) instead of a String in a future implementation.  This will throw compatibility with future user managers out the window if not adopted by the drftpd core, but it might resolve the above bug (assuming it works at all :D).

{{{
Merged from Offsite DrFTPD+:
Commit Date : 06/02/2006 - 11/30/06
Changeset : 227, 281, 317
Author : tdsoul
}}}